aiops-platform/topology-discovery/requirements.txt
neo4j==5.14.1
kafka-python-ng==2.2.3
python-dotenv==1.0.0
